+++ MatomoTracker/MatomoUserDefaults.swift
@@ -71,6 +71,16 @@ internal struct MatomoUserDefaults {
         }
     }
     
+    var forcedVisitorId: String? {
+        get {
+            return userDefaults.string(forKey: MatomoUserDefaults.Key.forcedVisitorID)
+        }
+        set {
+            userDefaults.setValue(newValue, forKey: MatomoUserDefaults.Key.forcedVisitorID)
+            userDefaults.synchronize()
+        }
+    }
+    
     var visitorUserId: String? {
         get {
             return userDefaults.string(forKey: MatomoUserDefaults.Key.visitorUserID);
@@ -99,6 +109,7 @@ extension MatomoUserDefaults {
         currentVisit = UserDefaults.standard.object(forKey: MatomoUserDefaults.Key.currentVisitTimestamp) as? Date
         optOut = UserDefaults.standard.bool(forKey: MatomoUserDefaults.Key.optOut)
         clientId = UserDefaults.standard.string(forKey: MatomoUserDefaults.Key.clientID)
+        forcedVisitorId = UserDefaults.standard.string(forKey: MatomoUserDefaults.Key.forcedVisitorID)
         visitorUserId = UserDefaults.standard.string(forKey: MatomoUserDefaults.Key.visitorUserID)
         lastOrder = UserDefaults.standard.object(forKey: MatomoUserDefaults.Key.lastOrder) as? Date
     }
@@ -114,6 +125,7 @@ extension MatomoUserDefaults {
         // Note:    To be compatible with previous versions, the clientID key retains its old value,
         //          even though it is now a misnomer since adding visitorUserID makes it a bit confusing.
         static let clientID = "PiwikVisitorIDKey"
+        static let forcedVisitorID = "PiwikForcedVisitorIDKey"
         static let visitorUserID = "PiwikVisitorUserIDKey"
         static let optOut = "PiwikOptOutKey"
         static let lastOrder = "PiwikLastOrderDateKey"