+++ MatomoTrackerTests/Fixtures/MemoryQueueFixtures.swift
@@ -2,7 +2,7 @@
 
 struct EventFixture {
     static func event() -> Event {
-        let visitor = Visitor(id: "spec_visitor_id", userId: "spec_user_id")
+        let visitor = Visitor(id: "spec_visitor_id", forcedId: nil, userId: "spec_user_id")
         let session = Session(sessionsCount: 0, lastVisit: Date(), firstVisit: Date())
         return Event(siteId: "spec_1", uuid: NSUUID.init(), visitor: visitor, session: session, date: Date(), url: URL(string: "http://spec_url")!, actionName: ["spec_action"], language: "spec_language", isNewSession: true, referer: nil,
                      customVariables: [],